verify-mcp/pyproject.toml
[project]
name = "srefix-verify-mcp"
version = "0.1.0"
description = "Metric-name verifier for srefix-diagnosis manuals — diffs metric refs in agents/*.md against per-tech whitelists captured from real exporters"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "PyYAML>=6.0",
]

[project.scripts]
srefix-verify = "verify_mcp.server:run"
srefix-verify-corpus = "verify_mcp.cli:run"
srefix-fix = "verify_mcp.fix_cli:run"

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
verify_mcp = ["whitelists/*.json"]
